requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
executing==1.2.0
eyeD3==0.8.10
filelock==3.6.0
frozenlist==1.4.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
grpcio==1.30.2
h11==0.14.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
install==1.3.5
ipython==8.14.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpickle==3.0.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
more-itertools==8.10.0
multidict==6.0.4
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
numpy==1.25.0
oauthlib==3.2.0
onboard==1.4.1
openpyxl==3.1.2
packaging==21.3
PAM==0.4.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.39
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
pyee==8.2.2
pyelftools==0.27
PyExcelerate==0.10.0
Pygments==2.15.1
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyppeteer==1.0.2
pyRFC3339==1.1
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-gnupg==0.4.8
python-magic==0.4.24
python-telegram-bot==20.4
python-whois==0.8.0
python-whois-extended==0.7.12
python-xlib==0.29
pytz==2022.1
pyvis==0.3.2
pyxdg==0.27
PyYAML==5.4.1
repo==0.3.0
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.1
stack-data==0.6.2
systemd-python==234
telebot==0.0.5
telegram==0.0.1
terminator==2.1.1
tinycss2==1.1.1
tldextract==3.1.2
tqdm==4.65.0
traitlets==5.9.0
tzdata==2023.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
whois==0.9.27
xdg==5
xkit==0.0.0
xlrd==1.2.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==1.0.0
